REST: Add query parameters to filter patches by
This adds generic filtering support to the PatchworkViewSet as well as
useful filtering options for listing patches.

DRF has some filter capabilities, but they can really slow down the web
interface for large data sets and are a little complex. This approach is
pretty simple and works efficiently.

The "options" method is now overridden to show the user available
filters when supported.

Signed-off-by: Andy Doan <andy.doan@linaro.org>